deployment/requirements.txt
numpy
pandas
streamlit
matplotlib
tensorflow
Pillow
seaborn
xmltodict
os-sys==2.1.4
